Updated CreatePatientHistory
Now in we include in the HistoryRegistration the Relevant Appointment. So the request will have to give the appointment ID. Also a new fucntion created in the backend, its for getting all the appointments that are in the state Created and Respected. Thats because the user in the Creating new registration will have to select an appointment.